feat(enriched): allow unmatched columns when enriching table
A table should headers that match all the regexes defined in a figure's
REGEX_GRID/Headers in the given order, but it can have unmatched columns
that will be disregarded when enriching the figure.

Example: Given we expect the headers ["Bytes", "Description"] in a table,
the following header row will also be considered valid, ["foo", "Bytes",
"bar", "Description", "baz"], and only column indices 1 and 3 will be
considered when evaluating the rest of the table.

Signed-off-by: Nadja Brix Koch <n.koch@samsung.com>