fix(skiplist): change error msg for wrong number of expected matches
The error message just ended with "true", because the actual boolean for
`skip.matches != count` was inserted in the format string. As this bool
is also the condition for raising the error, it doesn't give much value
by itself. With this commit, the format string is changed to print the
two numbers separated with ` != `, so it's easier to see what the actual
and expected number of matched skip-figures are.

Signed-off-by: Nadja Brix Koch <n.koch@samsung.com>